Crosswords (crossword.py)
Given code for generate.py to work. Project was also provided with 3 structure.txt files and 3 words.txt files. The structure files had a few rows of the crossword structure with the same length and different lengths of words available using _ and # like so: 
#____#
____##
______

The text files had lists of words.