Nim.py
Has adjoining file called play.py with the below code:
from nim import train, play

ai = train(10000)
play(ai)


In the game Nim, we begin with some number of piles, each with some number of objects. Players take turns: on a player’s turn, the player removes any non-negative number of objects from any one non-empty pile. Whoever removes the last object loses.

There’s some simple strategy you might imagine for this game: if there’s only one pile and three objects left in it, and it’s your turn, your best bet is to remove two of those objects, leaving your opponent with the third and final object to remove. But if there are more piles, the strategy gets considerably more complicated. In this problem, I've built an AI to learn the strategy for this game through reinforcement learning. By playing against itself repeatedly and learning from experience, eventually my AI  learns which actions to take and which actions to avoid.

In particular, I use Q-learning for this project. In Q-learning, we try to learn a reward value (a number) for every (state, action) pair. An action that loses the game will have a reward of -1, an action that results in the other player losing the game will have a reward of 1, and an action that results in the game continuing has an immediate reward of 0, but will also have some future reward.

I set the training to (10) to win! It's impressive at (10000).